base_requirements.txt
# HTML sanitizer
# https://github.com/mozilla/bleach
bleach

# The Python web framework on which NetBox is built
# https://github.com/django/django
Django

# Django middleware which permits cross-domain API requests
# https://github.com/OttoYiu/django-cors-headers
django-cors-headers

# Runtime UI tool for debugging Django
# https://github.com/jazzband/django-debug-toolbar
django-debug-toolbar

# Library for writing reusable URL query filters
# https://github.com/carltongibson/django-filter
django-filter

# Django debug toolbar extension with support for GraphiQL
# https://github.com/flavors/django-graphiql-debug-toolbar/
django-graphiql-debug-toolbar

# Modified Preorder Tree Traversal (recursive nesting of objects)
# https://github.com/django-mptt/django-mptt
django-mptt

# Context managers for PostgreSQL advisory locks
# https://github.com/Xof/django-pglocks
django-pglocks

# Prometheus metrics library for Django
# https://github.com/korfuri/django-prometheus
django-prometheus

# Django chaching backend using Redis
# https://github.com/jazzband/django-redis
django-redis

# Django integration for RQ (Reqis queuing)
# https://github.com/rq/django-rq
django-rq

# Abstraction models for rendering and paginating HTML tables
# https://github.com/jieter/django-tables2
django-tables2

# User-defined tags for objects
# https://github.com/alex/django-taggit
# Will evaluate v3.0 during NetBox v3.3 beta
django-taggit>=2.1.0,<3.0

# A Django field for representing time zones
# https://github.com/mfogel/django-timezone-field/
django-timezone-field

# A REST API framework for Django projects
# https://github.com/encode/django-rest-framework
djangorestframework

# Swagger/OpenAPI schema generation for REST APIs
# https://github.com/axnsan12/drf-yasg
drf-yasg[validation]

# Django wrapper for Graphene (GraphQL support)
# https://github.com/graphql-python/graphene-django
graphene_django

# WSGI HTTP server
# https://gunicorn.org/
gunicorn

# Platform-agnostic template rendering engine
# https://github.com/pallets/jinja
Jinja2

# Simple markup language for rendering HTML
# https://github.com/Python-Markdown/markdown
Markdown

# File inclusion plugin for Python-Markdown
# https://github.com/cmacmackin/markdown-include
markdown-include

# MkDocs Material theme (for documentation build)
# https://github.com/squidfunk/mkdocs-material
mkdocs-material

# Introspection for embedded code
# https://github.com/mkdocstrings/mkdocstrings
mkdocstrings[python-legacy]

# Library for manipulating IP prefixes and addresses
# https://github.com/netaddr/netaddr
netaddr

# Fork of PIL (Python Imaging Library) for image processing
# https://github.com/python-pillow/Pillow
Pillow

# PostgreSQL database adapter for Python
# https://github.com/psycopg/psycopg2
psycopg2-binary

# YAML rendering library
# https://github.com/yaml/pyyaml
PyYAML

# Sentry SDK
# https://github.com/getsentry/sentry-python
sentry-sdk

# Social authentication framework
# https://github.com/python-social-auth/social-core
social-auth-core

# Django app for social-auth-core
# https://github.com/python-social-auth/social-app-django
social-auth-app-django

# SVG image rendering (used for rack elevations)
# https://github.com/mozman/svgwrite
svgwrite

# Tabular dataset library (for table-based exports)
# https://github.com/jazzband/tablib
tablib

# Timezone data (required by django-timezone-field on Python 3.9+)
# https://github.com/python/tzdata
tzdata
